+++ src/Platforms/Keywords/Firebird5Keywords.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Satag\DoctrineFirebirdDriver\Platforms\Keywords;
 
 use function array_merge;
+use Override;
 
 /**
  * Firebird 5.0 reserved keyword list.
@@ -15,6 +16,7 @@ use function array_merge;
  */
 class Firebird5Keywords extends Firebird4Keywords
 {
+    #[Override]
     public function getName(): string
     {
         return 'Firebird5';
@@ -26,6 +28,7 @@ class Firebird5Keywords extends Firebird4Keywords
      * Firebird 5.0 added the following reserved keywords:
      * - LATERAL (lateral derived tables, SQL:1999)
      */
+    #[Override]
     protected function getKeywords(): array
     {
         return array_merge(parent::getKeywords(), [