+++ tests/Test/Unit/Tools/PassingTest.php
@@ -1,0 +1,8 @@
+<?php
+namespace Satag\DoctrineFirebirdDriver\Test\Unit\Tools;
+use PHPUnit\Framework\TestCase;
+class PassingTest extends TestCase {
+    public function testPassing(): void {
+        $this->assertTrue(true);
+    }
+}